requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
Django==5.2.7
django-crispy-forms==2.5
django-phonenumber-field==8.3.0
flake8==7.3.0
mccabe==0.7.0
mysqlclient==2.2.7
phonenumbers==9.0.19
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-decouple==3.8
sqlparse==0.5.3
